feat(engine): 重构虚拟机引擎并添加库支持
- 重命名和重构多个引擎相关文件，提高代码可读性和一致性
- 在 BinaryHeader 中添加库信息支持
- 更新方法调用和异常处理机制
- 优化字符串处理和常量池管理
-调整类型定义和方法签名

+++ assembler/src/main/java/io/github/leo1024/otrvm/conf/FieldMeta.java
@@ -63,7 +63,7 @@ public abstract class FieldMeta implements ISerializable {
             case INT -> new Int(name, number.intValue());
             case LONG -> new ILong(name, number.longValue());
             case SHORT -> new IShort(name, number.shortValue());
-            case FLOAT -> new IFloat(name, number.floatValue());
+            case DOUBLE -> new IFloat(name, number.floatValue());
             case BYTE -> new IByte(name, number.byteValue());
             default -> throw new ParserException("Invalid type.");
         };
